requirements.txt
python-telegram-bot==20.8
torch
openai-whisper
ffmpeg-python
pydub
